+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # 0.2.0
 
-- in-place template rendering
-- progress bar
-- fix exclusive run
+- Support in-place template rendering
+- Support progress bar
+- Fix exclusive run
+- Add a formal command-line application
 
 # 0.1.0